smoothing-project/requirements.txt
click==8.0.4
Flask==2.0.3
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.0.1
six==1.16.0
Werkzeug==2.0.3
